+++ server/utils/testUtils.js
@@ -5,12 +5,18 @@ exports.createUsers = (number = 1) => {
   let users = [];
 
   for (let x = 0; x < number; x++) {
-    User.create({
+    const userData = {
       email: faker.internet.email(),
       name: faker.name.findName(),
       password: faker.internet.password()
-    }, (err, user) => {
-      users.push(user);
+    };
+
+    User.create(userData, (err, user) => {
+      if (!err) {
+        users.push({_id: user._id.toString(), ...userData});
+      } else {
+        console.log(err)
+      }
     });
   }
